Made a map-function of radio buttons
