--- code/src/components/RadioOptionOne.js
@@ -1,19 +1,0 @@
-import React, { useState } from 'react';
-
-const RadioOption = () => {
-    return (
-        <>
-        <label htmlFor="rose">Rose</label>
-                <input 
-                    name="flower"
-                    id="rose"
-                    type="radio"
-                    value=""
-                    onChange="" 
-                />
-    </>
-    )
-
-}
-
-export default RadioOption;
