+++ code/src/components/Summary.js
@@ -2,12 +2,12 @@
 import React from 'react'
 import Form from './Form'
 
-const Summary = (prop) => {
+const Summary = ({ name, food, flower }) => {
     return (
         <>
             <div>
                 <h1>Hurra, en summary</h1>
-                <h3>Your answers: Your name: {prop.name}. You're favorite food: {prop.food}.</h3>
+                <h3>Your answers: Your name: {name}. You're favorite food: {food} Your favorite flower is {flower}.</h3>
             </div>
         </>
     )
